Update be logic for edit/delete existing posts

+++ backend/routes/posts.js
@@ -36,7 +36,8 @@ router.post(
     const post = new Post({
       title: req.body.title,
       content: req.body.content,
-      imagePath: url + "/images/" + req.file.filename
+      imagePath: url + "/images/" + req.file.filename,
+      creator: req.userData.userId
     });
     post.save()
       .then(createdPost => {
@@ -51,7 +52,8 @@ router.post(
 });
 
 router.put(
-  "/:id", checkAuth,
+  "/:id",
+  checkAuth,
   multer({ storage: storage }).single("image"),
   (req, res, next) => {
     let imagePath = req.body.imagePath;
@@ -65,9 +67,13 @@ router.put(
       content: req.body.content,
       imagePath: imagePath
     });
-    Post.updateOne({_id: req.params.id}, post)
+    Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post)
       .then(result => {
-        res.status(200).json({ message: "Update successful!" });
+        if (result.nModified > 0) {
+          res.status(200).json({ message: "Update successful!" });
+        } else {
+          res.status(401).json({ message: "Not Authorized!" });
+        }
     });
 });
 
@@ -105,10 +111,14 @@ router.get("/:id", (req, res, next) => {
 })
 
 router.delete("/:id", checkAuth, (req, res, next) => {
-  Post.deleteOne({_id: req.params.id})
+  Post.deleteOne({ _id: req.params.id, creator: req.userData.userId })
     .then(result => {
-      console.log(result);
-      res.status(200).json({ message: "Post deleted!" });
+      if(result.nModified > 0) {
+        res.status(200).json({ message: "Post deleted!" });
+      } else {
+        res.status(401).json({ message: "Not Authorized!" });
+      }
+
     });
 })
